refactor: move Args to core/args
