+++ src/main/java/org/tron/common/application/CliApplication.java
@@ -14,11 +14,11 @@
  */
 package org.tron.common.application;
 
+import org.tron.core.config.args.Args;
 import org.tron.core.db.BlockStore;
 import org.tron.core.db.Manager;
 import org.tron.core.net.node.Node;
 import org.tron.core.peer.Peer;
-import org.tron.program.Args;
 
 public class CliApplication implements Application {
